+++ src/graph.h
@@ -20,31 +20,31 @@ class GraphBase
                          const char *value,
                          WT_CURSOR *metadata_cursor);  // ✅ check if pvt
     std::string get_metadata(std::string key, WT_CURSOR *metadata_cursor);  // ✅
-    virtual node get_node(int node_id) = 0;                   // ✅
-    virtual node get_random_node() = 0;                       // ✅
-    virtual void create_new_graph() = 0;                      // ✅
-    virtual void add_node(node to_insert) = 0;                // ✅
-    virtual bool has_node(int node_id) = 0;                   // ✅
-    virtual void delete_node(int node_id) = 0;                // ✅
-    virtual void add_edge(edge to_insert, bool is_bulk) = 0;  // ✅
-    virtual void delete_edge(int src_id, int dst_id) = 0;     // ✅
-    virtual edge get_edge(int src_id, int dst_id) = 0;        // ✅
+    virtual node get_node(int64_t node_id) = 0;                    // ✅
+    virtual node get_random_node() = 0;                            // ✅
+    virtual void create_new_graph() = 0;                           // ✅
+    virtual void add_node(node to_insert) = 0;                     // ✅
+    virtual bool has_node(int64_t node_id) = 0;                    // ✅
+    virtual void delete_node(int64_t node_id) = 0;                 // ✅
+    virtual void add_edge(edge to_insert, bool is_bulk) = 0;       // ✅
+    virtual void delete_edge(int64_t src_id, int64_t dst_id) = 0;  // ✅
+    virtual edge get_edge(int64_t src_id, int64_t dst_id) = 0;     // ✅
     // void update_edge(edge to_update); no need to implement.
     virtual std::vector<node> get_nodes() = 0;  // ✅
     virtual std::vector<edge> get_edges() = 0;  // ✅
-    virtual bool has_edge(int src_id, int dst_id) = 0;
+    virtual bool has_edge(int64_t src_id, int64_t dst_id) = 0;
 
-    virtual int get_out_degree(int node_id) = 0;               // ✅
-    virtual int get_in_degree(int node_id) = 0;                // ✅
-    virtual std::vector<edge> get_out_edges(int node_id) = 0;  // ✅
-    virtual std::vector<node> get_out_nodes(int node_id) = 0;  // ✅
-    virtual std::vector<edge> get_in_edges(int node_id) = 0;   // ✅
-    virtual std::vector<node> get_in_nodes(int node_id) = 0;   // ✅
+    virtual uint32_t get_out_degree(int64_t node_id) = 0;          // ✅
+    virtual uint32_t get_in_degree(int64_t node_id) = 0;           // ✅
+    virtual std::vector<edge> get_out_edges(int64_t node_id) = 0;  // ✅
+    virtual std::vector<node> get_out_nodes(int64_t node_id) = 0;  // ✅
+    virtual std::vector<edge> get_in_edges(int64_t node_id) = 0;   // ✅
+    virtual std::vector<node> get_in_nodes(int64_t node_id) = 0;   // ✅
     void close();
-    int get_num_nodes();
-    int get_num_edges();
-    void set_num_nodes(int num_nodes, WT_CURSOR *cursor);
-    void set_num_edges(int num_edges, WT_CURSOR *cursor);
+    uint64_t get_num_nodes();
+    uint64_t get_num_edges();
+    void set_num_nodes(uint64_t num_nodes, WT_CURSOR *cursor);
+    void set_num_edges(uint64_t num_edges, WT_CURSOR *cursor);
     virtual void make_indexes() = 0;
     std::string get_db_name() const { return opts.db_name; };
